Add parent and dependencies to the traced results
Parent is a string with the ID of the module, which depended
on the current one and thus loaded it.

Dependencies is an array of strings with IDs of modules, which
the current one depends on.

This is useful to build a hierarchical tree of dependencies in a project,
which consists of multiple bundles, and to check if all module bundles are
distinct.